Move bin and bash files to ~/.dotfiles
Ideally the executables would go in /usr/local/.., which is
outside the git repo. The bash segments may be better all in a
single .bashrc file, which may help visualize the size and amount
of custom commands. Certain commands could also be standalone
executables not contained in the .bashrc file.